enh(interfaces): nipype interfaces API refactor
Resolves: #46.
